refactoring line api
