+++ src/api/Line/hooks.ts
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/react-query";
+import { LineApi } from "./functions";
+
+export const useGetLineNotificationStatus = () => {
+  return useQuery({
+    queryKey: ["users", "get-line-notification-status"],
+    queryFn: LineApi.getLineNotificationStatus,
+  });
+};
