+++ src/api/Line/types.ts
@@ -1,0 +1,3 @@
+export type GetLineNotificationStatusResponse = {
+  is_notification: boolean;
+};
