+++ src/api/User/functions.ts
@@ -3,7 +3,6 @@ import {
   CheckLineBotFriendResponse,
   CheckLoginedResponse,
   CreateLineAuthUrlResponse,
-  GetLineNotificationStatusResponse,
   LoginRequest,
   RegistRequest,
 } from "./type";
@@ -46,21 +45,12 @@ const checkLineBotFriend = async (): Promise<CheckLineBotFriendResponse> => {
   return res.data;
 };
 
-const getLineNotificationStatus =
-  async (): Promise<GetLineNotificationStatusResponse> => {
-    const res = await axios.get(ENDPOINT + "/api/line/notifications", {
-      withCredentials: true,
-    });
-    return res.data;
-  };
-
 const UserApi = {
   regist,
   login,
   checkLogined,
   createLineAuthUrl,
   checkLineBotFriend,
-  getLineNotificationStatus,
 };
 
 export default UserApi;
