+++ src/api/User/type.ts
@@ -26,7 +26,3 @@ export type CreateLineAuthUrlResponse = {
 export type CheckLineBotFriendResponse = {
   friend_flag: boolean;
 };
-
-export type GetLineNotificationStatusResponse = {
-  is_notification: boolean;
-};
